paper_assistant/requirements.txt
lightrag-hku
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
